src/main/java/com/example/charfreq/util/Sort.java
package com.example.charfreq.util;

import org.springframework.stereotype.Component;

import java.util.Comparator;
import java.util.LinkedHashMap;
import java.util.Map;
import java.util.stream.Collectors;

@Component
public class Sort implements Sortable<Map<Character, Long>, Map<Character, Long>> {

    @Override
    public Map<Character, Long> sortByDesc(Map<Character, Long> map) {
        return map.entrySet().stream()
                .sorted(Map.Entry.comparingByValue(Comparator.reverseOrder()))
                .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue, (k, v) -> k, LinkedHashMap::new));
    }
}